refactor : 패키지 분리
- domain을 새부 패키지로 분리
