+++ src/main/java/me/lxxjn0/game/domain/baseball/Baseballs.java
@@ -1,4 +1,4 @@
-package me.lxxjn0.game.domain;
+package me.lxxjn0.game.domain.baseball;
 
 import static java.util.stream.Collectors.*;
