+++ src/main/java/me/lxxjn0/game/domain/numberGenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package me.lxxjn0.game.domain;
+package me.lxxjn0.game.domain.numberGenerator;
 
 import java.util.List;
