+++ src/test/java/me/lxxjn0/game/domain/numberGenerator/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package me.lxxjn0.game.domain;
+package me.lxxjn0.game.domain.numberGenerator;
 
 import static org.assertj.core.api.Assertions.*;
